fix verifyHeader failure
1) recreate miner to avoid state issues, subsequent blocks were not allowed to be recreated once there was a failure, by resetting miner object it will allow to create subsequent blocks
2) only store sys mapping if its valid string > 0 length, don't need or want to store an empty string/hash

+++ core/rawdb/accessors_chain.go
@@ -711,17 +711,21 @@ func DeleteNEVMMappings(db ethdb.KeyValueWriter, sysBlockhash string, nevmBlockh
 	if err := db.Delete(nevmToSysKey(nevmBlockhash)); err != nil {
 		log.Crit("Failed to delete NEVM to hash mapping", "err", err)
 	}
-	if err := db.Delete(sysToNEVMKey(sysBlockhash)); err != nil {
-		log.Crit("Failed to delete SYS to NEVM mapping", "err", err)
+	if len(sysBlockhash) > 0 {
+		if err := db.Delete(sysToNEVMKey(sysBlockhash)); err != nil {
+			log.Crit("Failed to delete SYS to NEVM mapping", "err", err)
+		}
 	}
 }
 
 func WriteNEVMMappings(db ethdb.KeyValueWriter, sysBlockhash string, nevmBlockhash common.Hash) {
-	key := sysToNEVMKey(sysBlockhash)
-	if err := db.Put(key, nevmBlockhash.Bytes()); err != nil {
-		log.Crit("Failed to store header", "err", err)
+	if len(sysBlockhash) > 0 {
+		key := sysToNEVMKey(sysBlockhash)
+		if err := db.Put(key, nevmBlockhash.Bytes()); err != nil {
+			log.Crit("Failed to store header", "err", err)
+		}
 	}
-	key = nevmToSysKey(nevmBlockhash)
+	key := nevmToSysKey(nevmBlockhash)
 	if err := db.Put(key, []byte{0}); err != nil {
 		log.Crit("Failed to store header", "err", err)
 	}
